feat: EmailService 인터페이스 및 구현체 추가
